ui: updated credentials tab
